pan/midgard: Switch constants to uint32
Storing constants as float doesn't make sense when we have integer
instructions; better to switch to be integer natively and coerce to/from
float rather than the opposite.

+++ src/panfrost/midgard/compiler.h
@@ -109,7 +109,7 @@ typedef struct midgard_instruction {
         bool precede_break;
 
         bool has_constants;
-        float constants[4];
+        uint32_t constants[4];
         uint16_t inline_constant;
         bool has_blend_constant;
